Move beneficiary resolvers into dedicated modules
